+++ MAI_SGL/pipeline/clipper.h
@@ -8,11 +8,11 @@ namespace mai
 class Clipper
 {
 public:
-	static void do_clip_space(uint32_t draw_mode, const std::vector<VsOutput>& primitives, std::vector<VsOutput>& outputs);
-	static bool cull_face(uint32_t front_face, uint32_t cull_face, 
+	static void do_clip_space(uint8_t draw_mode, const std::vector<VsOutput>& primitives, std::vector<VsOutput>& outputs);
+	static bool cull_face(uint8_t front_face, uint8_t cull_face, 
 		const VsOutput& v0, const VsOutput& v1, const VsOutput& v2);
 private:
-	static void sutherland_hodgman(uint32_t draw_mode, const std::vector<VsOutput>& primitive, std::vector<VsOutput>& outputs);
+	static void sutherland_hodgman(uint8_t draw_mode, const std::vector<VsOutput>& primitive, std::vector<VsOutput>& outputs);
 
 	static bool inside(const mai::vec4f& point, const mai::vec4f& plane);
